Add SPARQL resources.
Add resources table, schema, controller, view, and templates
to hold the prefix labels and the IRI's.

Add list of resources next to current query, with the capability
to add, modify, or delete.

Change the API to nest the resources within the queries.
The the query id will be sent to the resources controller, so it
will be able to redirect to the user's current page after adding,
modifying, or deleting a resource.